Don't return a value from the response handler.
When the clack app uses a delayed response and returns a NIL body,
HANDLE-RESPONSE has not called SEND-HEADERS, so hunchentoot expects the
return value of the handler to be the body of the response. As it is,
the handler is returning the value of the delayed response function,
which is usually not something we want sent. Instead, make sure the
handler never returns a value, since HANDLE-NORMAL-RESPONSE should
always be taking care of that.