Correctly handle prematurely closed sockets in Hunchentoot handler
Without this change, a condition exists wherein a client that closes a
connection prematurely will cause a TYPE-ERROR condition when the
Hunchentoot handler tries to write the 500 error to the now-NIL socket
when :debug is false, with the result that (at least in SBCL) the worker
thread becomes stuck. This change logs the error but allows normal
processing to continue.